requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
pytesseract
Pillow
alembic
python-multipart
PyJWT